💥 Use string error codes
This breaking change moves us from number error codes to strings to keep
better-aligned with `sharedb`.

We also move error classes into their own module so that clients can
check for them using `instanceof`.
